requirements.txt
fastapi
uvicorn
pillow
numpy
torch
torchvision
python-multipart